Moved deposit sweep proof submission into a separate package
